+++ src/app/(site)/_components/nav-actions.tsx
@@ -1,10 +1,37 @@
 "use client";
 
-import React from "react";
+import React, { useState } from "react";
 
 import Button from "@/app/_components/ui/button";
+import Modal from "@/app/_components/ui/modal";
+import GetInTouchForm from "./get-in-touch-form";
+import SectionTitle from "./section-title";
+import { Typography } from "@/app/_components/ui/typography";
 
 export default function NavActions() {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const onFreeConsultationClick = () => {
+    const getInTouchContainer = document.querySelector(
+      "#get-in-touch-container"
+    );
+    console.log(getInTouchContainer);
+    setTimeout(() => {
+      if (getInTouchContainer) {
+        getInTouchContainer.scrollIntoView({
+          behavior: "smooth",
+          block: "center",
+        });
+      } else {
+        openModel();
+      }
+    }, 100);
+  };
+
+  const openModel = () => {
+    setIsOpen(true);
+  };
+
   return (
     <div className="flex flex-col sm:flex-row gap-5">
       <Button
@@ -20,9 +47,19 @@ export default function NavActions() {
         size="md"
         customClass="w-fit"
         onClick={() => {
-          alert("");
+          onFreeConsultationClick();
         }}
       />
+
+      <Modal
+        isOpen={isOpen}
+        onClose={() => {
+          setIsOpen(false);
+        }}
+      >
+        <Typography variant="h3" className="my-5 mx-auto text-center">Get in Touch</Typography>
+        <GetInTouchForm />
+      </Modal>
     </div>
   );
 }
